--- a/src/com/xilinx/rapidwright/interchange/DeviceResourcesWriter.java
+++ b/src/com/xilinx/rapidwright/interchange/DeviceResourcesWriter.java
@@ -721,8 +721,7 @@ public class DeviceResourcesWriter {
 
     public static void writeAllWiresAndNodesToBuilder(Device device, DeviceResources.Device.Builder devBuilder) {
         LongEnumerator allWires = new LongEnumerator();
-        LongEnumerator allNodes = new LongEnumerator();
-
+        ArrayList<Long> allNodes = new ArrayList();
 
         for(Tile tile : device.getAllTiles()) {
             for(int i=0; i < tile.getWireCount(); i++) {
@@ -732,8 +731,8 @@ public class DeviceResourcesWriter {
                 Node node = wire.getNode();
                 if(node == null)
                     continue;
-                if (node.getTile() == tile)
-                    allNodes.addObject(makeKey(node.getTile(), node.getWire()));
+                if (node.getTile() == tile && node.getWire() == i)
+                    allNodes.add(makeKey(node.getTile(), node.getWire()));
             }
         }