@clavin-xlnx 2021-05-19
Thanks! This optimization looks good to me.